Move adlr-cli tool up to top-level dir since there is only one cmd
